+++ books/templates/partials/_recommendation.html
@@ -1,6 +1,7 @@
 {% load custom_filters %}
 
-<div class="flex flex-col">
+<div class="flex flex-col" id="recommendation-{{book.id}}" hx-target="#recommendation-{{book.id}}" hx-swap="outerHTML"
+    hx-headers='{"X-CSRFToken": "{{ csrf_token }}"}'>
     <a href="{% url 'book_details' book.id %}">
         <p class="text-4xl font-bold">{{ book.title }}<span class="text-sm font-light"> by
                 {{ book.author|format_authors}}</span>
@@ -22,7 +23,7 @@
         <div class="flex flex-row gap-1">
             {% if request.user in book.likes.all %}
             <!-- Thumbs Up Solid -->
-            <a href="#" class="btn-sm">
+            <a hx-post="{% url 'like' book.id %}" class="btn-sm">
                 <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" fill="currentColor" class="size-6">
                     <path
                         d="M7.493 18.5c-.425 0-.82-.236-.975-.632A7.48 7.48 0 0 1 6 15.125c0-1.75.599-3.358 1.602-4.634.151-.192.373-.309.6-.397.473-.183.89-.514 1.212-.924a9.042 9.042 0 0 1 2.861-2.4c.723-.384 1.35-.956 1.653-1.715a4.498 4.498 0 0 0 .322-1.672V2.75A.75.75 0 0 1 15 2a2.25 2.25 0 0 1 2.25 2.25c0 1.152-.26 2.243-.723 3.218-.266.558.107 1.282.725 1.282h3.126c1.026 0 1.945.694 2.054 1.715.045.422.068.85.068 1.285a11.95 11.95 0 0 1-2.649 7.521c-.388.482-.987.729-1.605.729H14.23c-.483 0-.964-.078-1.423-.23l-3.114-1.04a4.501 4.501 0 0 0-1.423-.23h-.777ZM2.331 10.727a11.969 11.969 0 0 0-.831 4.398 12 12 0 0 0 .52 3.507C2.28 19.482 3.105 20 3.994 20H4.9c.445 0 .72-.498.523-.898a8.963 8.963 0 0 1-.924-3.977c0-1.708.476-3.305 1.302-4.666.245-.403-.028-.959-.5-.959H4.25c-.832 0-1.612.453-1.918 1.227Z" />
@@ -30,7 +31,7 @@
             </a>
             {% else %}
             <!-- Thumbs Up Outlined -->
-            <a href="#" class="btn-sm">
+            <a hx-post="{% url 'like' book.id %}" class="btn-sm">
                 <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5"
                     stroke="currentColor" class="size-6">
                     <path stroke-linecap="round" stroke-linejoin="round"
@@ -42,7 +43,7 @@
         </div>
         {% if request.user in book.favorites.all %}
         <!-- Bookmark Solid -->
-        <a href="#" class="btn-sm">
+        <a hx-post="{% url 'favorite' book.id %}" class="btn-sm">
             <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" fill="currentColor" class="size-6">
                 <path fill-rule="evenodd"
                     d="M6.32 2.577a49.255 49.255 0 0 1 11.36 0c1.497.174 2.57 1.46 2.57 2.93V21a.75.75 0 0 1-1.085.67L12 18.089l-7.165 3.583A.75.75 0 0 1 3.75 21V5.507c0-1.47 1.073-2.756 2.57-2.93Z"
@@ -51,7 +52,7 @@
         </a>
         {% else %}
         <!-- Bookmark Outlined -->
-        <a href="#" class="btn-sm">
+        <a hx-post="{% url 'favorite' book.id %}" class="btn-sm">
             <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5"
                 stroke="currentColor" class="size-6">
                 <path stroke-linecap="round" stroke-linejoin="round"
