+++ books/urls.py
@@ -5,4 +5,6 @@ urlpatterns = [
     path("recommendations/", views.recommendations, name="recommendations"),
     path("submissions/", views.submissions, name="submissions"),
     path("details/<int:id>/", views.details, name="book_details"),
+    path("favorite/<int:id>/", views.favorite_book, name="favorite"),
+    path("like/<int:id>/", views.like_book, name="like"),
 ]
