+++ templates/base.html
@@ -5,13 +5,14 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    {% if description %}
+    <meta name="description" content="{% block description %}{% endblock description %}">
+    {% endif %}
     <link rel="icon" type="image/x-icon" href="{% static '/img/favicon.ico' %}">
     <link rel="apple-touch-icon" sizes="180x180" href="{% static '/img/apple-touch-icon.png' %}">
     <link rel="icon" type="image/png" sizes="32x32" href="{% static '/img/favicon-32x32.png' %}">
     <link rel="icon" type="image/png" sizes="16x16" href="{% static '/img/favicon-16x16.png' %}">
-    <script src="https://unpkg.com/htmx.org@2.0.3"
-        integrity="sha384-0895/pl2MU10Hqc6jd4RvrthNlDiE9U1tWmX7WRESftEDRosgxNsQG/Ze9YMRzHq"
-        crossorigin="anonymous"></script>
+    <script src="{% static 'js/htmx/htmx.min.js' %}" defer></script>
     {% tailwind_css %}
     <title>{% block title %}{% endblock %}</title>
 </head>
